Enum juniper::InputValue

pub enum InputValue {
    Null,
    Int(i32),
    Float(f64),
    String(String),
    Boolean(bool),
    Enum(String),
    Variable(String),
    List(Vec<Spanning<InputValue>>),
    Object(Vec<(Spanning<String>, Spanning<InputValue>)>),
}

A JSON-like value that can be passed into the query execution, either out-of-band, or in-band as default variable values. These are not constant and might contain variables.

Lists and objects variants are spanned, i.e. they contain a reference to their position in the source file, if available.

Variants
NullInt(i32)Float(f64)String(String)Boolean(bool)Enum(String)Variable(String)List(Vec<Spanning<InputValue>>)Object(Vec<(Spanning<String>, Spanning<InputValue>)>)
Methods

impl InputValue

pub fn null() -> InputValue

Construct a null value.

pub fn int(i: i32) -> InputValue

Construct an integer value.

pub fn float(f: f64) -> InputValue

Construct a floating point value.

pub fn boolean(b: bool) -> InputValue

Construct a boolean value.

pub fn string<T: AsRef<str>>(s: T) -> InputValue

Construct a string value.

pub fn enum_value<T: AsRef<str>>(s: T) -> InputValue

Construct an enum value.

pub fn variable<T: AsRef<str>>(v: T) -> InputValue

Construct a variable value.

pub fn list(l: Vec<InputValue>) -> InputValue

Construct an unlocated list.

Convenience function to make each InputValue in the input vector not contain any location information. Can be used from ToInputValue implementations, where no source code position information is available.

pub fn parsed_list(l: Vec<Spanning<InputValue>>) -> InputValue

Construct a located list.

pub fn object<K>(o: OrderMap<K, InputValue>) -> InputValue where
K: AsRef<str> + Eq + Hash,

Construct an unlocated object.

Similar to InputValue::list, it makes each key and value in the given hash map not contain any location information.

pub fn parsed_object(
o: Vec<(Spanning<String>, Spanning<InputValue>)>
) -> InputValue

Construct a located object.

pub fn into_const(self, vars: &Variables) -> InputValue

Resolve all variables to their values.

pub fn convert<T>(&self) -> Option<T> where
T: FromInputValue,

Shorthand form of invoking FromInputValue::from().

pub fn is_null(&self) -> bool

Does the value represent null?

pub fn is_variable(&self) -> bool

Does the value represent a variable?

pub fn as_enum_value(&self) -> Option<&str>

View the underlying enum value, if present.

pub fn as_int_value(&self) -> Option<i32>

View the underlying int value, if present.

pub fn as_float_value(&self) -> Option<f64>

View the underlying float value, if present.

pub fn as_string_value(&self) -> Option<&str>

View the underlying string value, if present.

pub fn to_object_value(&self) -> Option<OrderMap<&str, &InputValue>>

Convert the input value to an unlocated object value.

This constructs a new OrderMap that contain references to the keys and values in self.

pub fn to_list_value(&self) -> Option<Vec<&InputValue>>

Convert the input value to an unlocated list value.

This constructs a new vector that contain references to the values in self.

pub fn referenced_variables(&self) -> Vec<&str>

Recursively find all variables

pub fn unlocated_eq(&self, other: &InputValue) -> bool

Compare equality with another InputValue ignoring any source position information.