Enum juniper::Value

pub enum Value {
    Null,
    Int(i32),
    Float(f64),
    String(String),
    Boolean(bool),
    List(Vec<Value>),
    Object(OrderMap<String, Value>),
}

Serializable value returned from query and field execution.

Used by the execution engine and resolvers to build up the response structure. Similar to the Json type found in the serialize crate.

It is also similar to the InputValue type, but can not contain enum values or variables. Also, lists and objects do not contain any location information since they are generated by resolving fields and values rather than parsing a source query.

Variants
NullInt(i32)Float(f64)String(String)Boolean(bool)List(Vec<Value>)Object(OrderMap<String, Value>)
Methods

impl Value

pub fn null() -> Value

Construct a null value.

pub fn int(i: i32) -> Value

Construct an integer value.

pub fn float(f: f64) -> Value

Construct a floating point value.

pub fn string<T: AsRef<str>>(s: T) -> Value

Construct a string value.

pub fn boolean(b: bool) -> Value

Construct a boolean value.

pub fn list(l: Vec<Value>) -> Value

Construct a list value.

pub fn object<K>(o: OrderMap<K, Value>) -> Value where
K: Into<String> + Eq + Hash,

Construct an object value.

pub fn is_null(&self) -> bool

Does this value represent null?

pub fn as_float_value(&self) -> Option<&f64>

View the underlying float value, if present.

pub fn as_object_value(&self) -> Option<&OrderMap<String, Value>>

View the underlying object value, if present.

pub fn as_mut_object_value(&mut self) -> Option<&mut OrderMap<String, Value>>

Mutable view into the underlying object value, if present.

pub fn as_list_value(&self) -> Option<&Vec<Value>>

View the underlying list value, if present.

pub fn as_string_value(&self) -> Option<&str>

View the underlying string value, if present.