Struct juniper::iron_handlers::GraphQLHandler

pub struct GraphQLHandler<CtxFactory, Query, Mutation, CtxT> where CtxFactory: Fn(&mut Request) -> CtxT + Send + Sync + 'static, CtxT: Send + Sync + 'static, Query: GraphQLType<CtxT> + Send + Sync + 'static, Mutation: GraphQLType<CtxT> + Send + Sync + 'static {
    // some fields omitted
}

Handler that executes GraphQL queries in the given schema

The handler responds to GET requests and POST requests only. In GET requests, the query should be supplied in the query URL parameter, e.g. http://localhost:3000/graphql?query={hero{name}}.

POST requests support both queries and variables. POST a JSON document to this endpoint containing the field "query" and optionally "variables". The variables should be a JSON object containing the variable to value mapping.

Methods

impl<CtxFactory, Query, Mutation, CtxT> GraphQLHandler<CtxFactory, Query, Mutation, CtxT> where CtxFactory: Fn(&mut Request) -> CtxT + Send + Sync + 'static, CtxT: Send + Sync + 'static, Query: GraphQLType<CtxT> + Send + Sync + 'static, Mutation: GraphQLType<CtxT> + Send + Sync + 'static

fn new(context_factory: CtxFactory, query: Query, mutation: Mutation) -> Self

Build a new GraphQL handler

The context factory will receive the Iron request object and is expected to construct a context object for the given schema. This can be used to construct e.g. database connections or similar data that the schema needs to execute the query.