Struct juniper::meta::Argument

pub struct Argument {
    // some fields omitted
}

Metadata for an argument to a field

Methods

impl Argument

fn description(self, description: &str) -> Argument

Set the description of the argument

This overwrites the description if any was previously set.

fn default_value(self, default_value: InputValue) -> Argument

Set the default value of the argument

This overwrites the description if any was previously set.