Struct juniper::meta::EnumValue

pub struct EnumValue {
    pub name: String,
    pub description: Option<String>,
    pub deprecation_reason: Option<String>,
}

Metadata for a single value in an enum

Fields
name: String

The name of the enum value

This is the string literal representation of the enum in responses.

description: Option<String>

The optional description of the enum value.

Note: this is not the description of the enum itself; it's the description of this enum value.

deprecation_reason: Option<String>

The optional deprecation reason

If this is Some, the field will be considered isDeprecated.

Methods

impl EnumValue

pub fn new(name: &str) -> EnumValue

Construct a new enum value with the provided name

pub fn description(self, description: &str) -> EnumValue

Set the description of the enum value

This overwrites the description if any was previously set.

pub fn deprecated(self, reason: &str) -> EnumValue

Set the deprecation reason for the enum value

This overwrites the deprecation reason if any was previously set.