Enum juniper::parser::LexerError

pub enum LexerError {
    UnknownCharacter(char),
    UnexpectedCharacter(char),
    UnterminatedString,
    UnknownCharacterInString(char),
    UnknownEscapeSequence(String),
    UnexpectedEndOfFile,
    InvalidNumber,
}

Error when tokenizing the input source

Variants
UnknownCharacter(char)

An unknown character was found

Unknown characters are characters that do not occur anywhere in the GraphQL language, such as ? or %.

UnexpectedCharacter(char)

An unexpected character was found

Unexpected characters are characters that do exist in the GraphQL language, but is not expected at the current position in the document.

UnterminatedString

An unterminated string literal was found

Apart from forgetting the ending ", terminating a string within a Unicode escape sequence or having a line break in the string also causes this error.

UnknownCharacterInString(char)

An unknown character in a string literal was found

This occurs when an invalid source character is found in a string literal, such as ASCII control characters.

UnknownEscapeSequence(String)

An unknown escape sequence in a string literal was found

Only a limited set of escape sequences are supported, this is emitted when e.g. "\l" is parsed.

UnexpectedEndOfFile

The input source was unexpectedly terminated

Emitted when the current token requires a succeeding character, but the source has reached EOF. Emitted when scanning e.g. "1.".

InvalidNumber

An invalid number literal was found