Struct juniper::parser::Spanning

pub struct Spanning<T: Debug> {
    pub item: T,
    pub start: SourcePosition,
    pub end: SourcePosition,
}

Data structure used to wrap items with start and end markers in the input source

A "span" is a range of characters in the input source, starting at the character pointed by the start field and ending just before the end marker.

Fields
item: T

The wrapped item

start: SourcePosition

Start position of the item

end: SourcePosition

End position of the item

This points to the first source position after the wrapped item.

Methods

impl<T: Debug> Spanning<T>

fn map<O: Debug, F: Fn(T) -> O>(self, f: F) -> Spanning<O>

Modify the contents of the spanned item