Struct juniper::Arguments

pub struct Arguments {
    // some fields omitted
}

Field argument container

Methods

impl Arguments

fn get<T>(&self, key: &str) -> Option<T> where T: FromInputValue

Get and convert an argument into the desired type.

If the argument is found, or a default argument has been provided, the InputValue will be converted into the type T.

Returns Some if the argument is present and type conversion succeeeds.