Struct juniper::ExecutionError

pub struct ExecutionError {
    // some fields omitted
}

Error type for errors that occur during query execution

All execution errors contain the source position in the query of the field that failed to resolve. It also contains the field stack.

Methods

impl ExecutionError

fn message(&self) -> &str

The error message

fn location(&self) -> &SourcePosition

The source location in the query of the field that failed to resolve

fn path(&self) -> &[String]

The path of fields leading to the field that generated this error