Struct juniper::Object

pub struct Object<S> { /* fields omitted */ }

A Object value

Methods

impl<S> Object<S>

pub fn with_capacity(size: usize) -> Self

Create a new Object value with a fixed number of preallocated slots for field-value pairs

pub fn add_field<K>(&mut self, k: K, value: Value<S>) -> Option<Value<S>> where
K: Into<String>,
&'a str: PartialEq<K>,

Add a new field with a value

If there is already a field with the same name the old value is returned

pub fn contains_field<K>(&self, f: K) -> bool where
&'a str: PartialEq<K>,

Check if the object already contains a field with the given name

pub fn iter(&self) -> impl Iterator<Item = &(String, Value<S>)>

Get a iterator over all field value pairs

pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut (String, Value<S>)>

Get a iterator over all mutable field value pairs

pub fn field_count(&self) -> usize

Get the current number of fields

pub fn get_field_value<K>(&self, key: K) -> Option<&Value<S>> where
&'a str: PartialEq<K>,

Get the value for a given field

pub fn sort_by_field(&mut self)

Recursively sort all keys by field.