Trait juniper::FromInputValue

pub trait FromInputValue: Sized {
    fn from(v: &InputValue) -> Option<Self>;
}

Parse an unstructured input value into a Rust data type.

The conversion can fail, and must in that case return None. Implemented automatically by the convenience macros graphql_enum! and graphql_scalar!. Must be implemented manually when manually exposing new enums or scalars.

Required Methods

fn from(v: &InputValue) -> Option<Self>

Performs the conversion.