Struct juniper::validation::RuleError

pub struct RuleError {
    // some fields omitted
}

Query validation error

Methods

impl RuleError

fn message(&self) -> &str

Access the message for a validation error

fn locations(&self) -> &[SourcePosition]

Access the positions of the validation error

All validation errors contain at least one source position, but some validators supply extra context through multiple positions.