Crate juniper

GraphQL

GraphQL is a data query language developed by Facebook intended to serve mobile and web application frontends.

Juniper makes it possible to write GraphQL servers in Rust that are type-safe and blazingly fast. We also try to make declaring and resolving GraphQL schemas as convenient as possible as Rust will allow.

Juniper does not include a web server - instead it provides building blocks to make integration with existing servers straightforward. It optionally provides a pre-built integration for the Iron and Rocket frameworks, including embedded Graphiql for easy debugging.

Cargo crate

API Reference

Book: Guides and Examples

Getting Started

The best place to get started is the Juniper Book, which contains guides with plenty of examples, covering all features of Juniper.

To get started quickly and get a feel for Juniper, check out the Quickstart section.

For specific information about macros, types and the Juniper api, the API Reference is the best place to look.

You can also check out src/tests/schema.rs to see a complex schema including polymorphism with traits and interfaces. For an example of web framework integration, see the rocket and iron examples folders.

Features

Juniper supports the full GraphQL query language according to the specification, including interfaces, unions, schema introspection, and validations. It does not, however, support the schema language.

As an exception to other GraphQL libraries for other languages, Juniper builds non-null types by default. A field of type Vec<Episode> will be converted into [Episode!]!. The corresponding Rust type for e.g. [Episode] would be Option<Vec<Option<Episode>>>.

Integrations

Data types

Juniper has automatic integration with some very common Rust crates to make building schemas a breeze. The types from these crates will be usable in your Schemas automatically.

uuid

url

chrono

Web Frameworks

rocket

iron

API Stability

Juniper has not reached 1.0 yet, thus some API instability should be expected.

Re-exports

pub use http::graphiql;

Modules

http
Utilities for building HTTP endpoints in a library-agnostic manner

integrations
Provides GraphQLType implementations for some external types

meta
Types used to describe a GraphQL schema

parser
Query parser and language utilities

Macros

graphql_interface
Expose GraphQL interfaces

graphql_object
Expose GraphQL objects

graphql_scalar
Expose GraphQL scalars

graphql_union
Expose GraphQL unions

graphql_value
Construct JSON-like values by using JSON syntax

Structs

Arguments
Field argument container

EmptyMutation
Utility type to define read-only schemas

ExecutionError
Error type for errors that occur during query execution

Executor
Query execution engine

FieldError
Error type for errors that occur during field resolution

ID
An ID as defined by the GraphQL specification

LookAheadArgument
An argument passed into the query

LookAheadSelection
A selection performed by a query

Object
A Object value

Registry
A type registry used to build schemas

RootNode
Root query node of a schema

RuleError
Query validation error

Enums

Applies
An enum that describes if a field is available in all types of the interface or only in a certain subtype

DefaultScalarValue
The default scalar value representation in juniper

GraphQLError
An error that prevented query execution

InputValue
A JSON-like value that can be passed into the query execution, either out-of-band, or in-band as default variable values. These are not constant and might contain variables.

LookAheadValue
A JSON-like value that can is used as argument in the query execution

Selection
Entry in a GraphQL selection set

Type
A type literal in the syntax tree

TypeKind
GraphQL type kind

Value
Serializable value returned from query and field execution.

Traits

Context
Marker trait for types that can act as context objects for GraphQL types.

FromContext
Conversion trait for context types

FromInputValue
Parse an unstructured input value into a Rust data type.

GraphQLType
Primary trait used to expose Rust types in a GraphQL schema

IntoFieldError
Custom error handling trait to enable Error types other than FieldError to be specified as return value.

LookAheadMethods
A set of common methods for ConcreteLookAheadSelection and LookAheadSelection

ParseScalarValue
A trait used to convert a ScalarToken into a certain scalar value type

ScalarRefValue
A marker trait extending the ScalarValue trait

ScalarValue
A trait marking a type that could be used as internal representation of scalar values in juniper

ToInputValue
Losslessly clones a Rust data type into an InputValue.

Functions

execute
Execute a query in a provided schema

Type Definitions

ExecutionResult
The result of resolving an unspecified field

FieldResult
The result of resolving the value of a field of type T

ParseScalarResult
The result of converting a string into a scalar value

Variables
The map of variables used for substitution during query execution