Enum juniper::InputValue

pub enum InputValue<S = DefaultScalarValue> {
    Null,
    Scalar(S),
    Enum(String),
    Variable(String),
    List(Vec<Spanning<InputValue<S>>>),
    Object(Vec<(Spanning<String>, Spanning<InputValue<S>>)>),
}

A JSON-like value that can be passed into the query execution, either out-of-band, or in-band as default variable values. These are not constant and might contain variables.

Lists and objects variants are spanned, i.e. they contain a reference to their position in the source file, if available.

Variants
NullScalar(S)Enum(String)Variable(String)List(Vec<Spanning<InputValue<S>>>)Object(Vec<(Spanning<String>, Spanning<InputValue<S>>)>)
Methods

impl<S> InputValue<S> where
S: ScalarValue,

pub fn null() -> Self

Construct a null value.

pub fn int(i: i32) -> Self

Deprecated since 0.11.0:
Use InputValue::scalar instead

Construct an integer value.

pub fn float(f: f64) -> Self

Deprecated since 0.11.0:
Use InputValue::scalar instead

Construct a floating point value.

pub fn boolean(b: bool) -> Self

Deprecated since 0.11.0:
Use InputValue::scalar instead

Construct a boolean value.

pub fn string<T: AsRef<str>>(s: T) -> Self

Deprecated since 0.11.0:
Use InputValue::scalar instead

Construct a string value.

pub fn scalar<T>(v: T) -> Self where
T: Into<S>,

Construct a scalar value

pub fn enum_value<T: AsRef<str>>(s: T) -> Self

Construct an enum value.

pub fn variable<T: AsRef<str>>(v: T) -> Self

Construct a variable value.

pub fn list(l: Vec<Self>) -> Self

Construct an unlocated list.

Convenience function to make each InputValue in the input vector not contain any location information. Can be used from ToInputValue implementations, where no source code position information is available.

pub fn parsed_list(l: Vec<Spanning<Self>>) -> Self

Construct a located list.

pub fn object<K>(o: IndexMap<K, Self>) -> Self where
K: AsRef<str> + Eq + Hash,

Construct an unlocated object.

Similar to InputValue::list, it makes each key and value in the given hash map not contain any location information.

pub fn parsed_object(o: Vec<(Spanning<String>, Spanning<Self>)>) -> Self

Construct a located object.

pub fn into_const(self, vars: &Variables<S>) -> Self

Resolve all variables to their values.

pub fn convert<T>(&self) -> Option<T> where
T: FromInputValue<S>,
&'b S: ScalarRefValue<'b>,

Shorthand form of invoking FromInputValue::from().

pub fn is_null(&self) -> bool

Does the value represent null?

pub fn is_variable(&self) -> bool

Does the value represent a variable?

pub fn as_enum_value(&self) -> Option<&str>

View the underlying enum value, if present.

pub fn as_int_value<'a>(&'a self) -> Option<i32> where
&'a S: Into<Option<&'a i32>>,

Deprecated since 0.11.0:
Use InputValue::as_scalar_value instead

View the underlying int value, if present.

pub fn as_float_value<'a>(&'a self) -> Option<f64> where
&'a S: Into<Option<&'a f64>>,

Deprecated since 0.11.0:
Use InputValue::as_scalar_value instead

View the underlying float value, if present.

pub fn as_string_value<'a>(&'a self) -> Option<&'a str> where
&'a S: Into<Option<&'a String>>,

Deprecated since 0.11.0:
Use InputValue::as_scalar_value instead

View the underlying string value, if present.

pub fn as_scalar(&self) -> Option<&S>

View the underlying scalar value, if present.

pub fn as_scalar_value<'a, T>(&'a self) -> Option<&'a T> where
T: 'a,
&'a S: Into<Option<&'a T>>,

View the underlying scalar value, if present.

pub fn to_object_value<'a>(&'a self) -> Option<IndexMap<&'a str, &'a Self>>

Convert the input value to an unlocated object value.

This constructs a new IndexMap that contain references to the keys and values in self.

pub fn to_list_value(&self) -> Option<Vec<&Self>>

Convert the input value to an unlocated list value.

This constructs a new vector that contain references to the values in self.

pub fn referenced_variables(&self) -> Vec<&str>

Recursively find all variables

pub fn unlocated_eq(&self, other: &Self) -> bool

Compare equality with another InputValue ignoring any source position information.