Enum juniper::TypeKind

pub enum TypeKind {
    Scalar,
    Object,
    Interface,
    Union,
    Enum,
    InputObject,
    List,
    NonNull,
}

GraphQL type kind

The GraphQL specification defines a number of type kinds - the meta type
of a type.

Variants

Scalar

Scalar types

Scalar types appear as the leaf nodes of GraphQL queries. Strings,
numbers, and booleans are the built in types, and while it's possible
to define your own, it's relatively uncommon.

Object

Object types

The most common type to be implemented by users. Objects have fields
and can implement interfaces.

Interface

Interface types

Interface types are used to represent overlapping fields between
multiple types, and can be queried for their concrete type.

Union

Union types

Unions are similar to interfaces but can not contain any fields on
their own.

Enum

Enum types

Like scalars, enum types appear as the leaf nodes of GraphQL queries.

InputObject

Input objects

Represents complex values provided in queries into the system.

List

List types

Represent lists of other types. This library provides implementations
for vectors and slices, but other Rust types can be extended to serve
as GraphQL lists.

NonNull

Non-null types

In GraphQL, nullable types are the default. By putting a ! after a
type, it becomes non-nullable.