Enum juniper::Value

pub enum Value {
    Null,
    Int(i64),
    Float(f64),
    String(String),
    Boolean(bool),
    List(Vec<Value>),
    Object(HashMap<String, Value>),
}

Serializable value returned from query and field execution.

Used by the execution engine and resolvers to build up the response structure. Similar to the Json type found in the serialize crate.

It is also similar to the InputValue type, but can not contain enum values or variables. Also, lists and objects do not contain any location information since they are generated by resolving fields and values rather than parsing a source query.

Variants
NullInt(i64)Float(f64)String(String)Boolean(bool)List(Vec<Value>)Object(HashMap<String, Value>)
Methods

impl Value

fn null() -> Value

Construct a null value.

fn int(i: i64) -> Value

Construct an integer value.

fn float(f: f64) -> Value

Construct a floating point value.

fn string<T: AsRef<str>>(s: T) -> Value

Construct a string value.

fn boolean(b: bool) -> Value

Construct a boolean value.

fn list(l: Vec<Value>) -> Value

Construct a list value.

fn object<K>(o: HashMap<K, Value>) -> Value where K: AsRef<str> + Eq + Hash

Construct an object value.

fn is_null(&self) -> bool

Does this value represent null?

fn as_object_value(&self) -> Option<&HashMap<String, Value>>

View the underlying object value, if present.

fn as_list_value(&self) -> Option<&Vec<Value>>

View the underlying list value, if present.

fn as_string_value(&self) -> Option<&str>

View the underlying string value, if present.