Enum juniper::Value

pub enum Value<S = DefaultScalarValue> {
    Null,
    Scalar(S),
    List(Vec<Value<S>>),
    Object(Object<S>),
}

Serializable value returned from query and field execution.

Used by the execution engine and resolvers to build up the response structure. Similar to the Json type found in the serialize crate.

It is also similar to the InputValue type, but can not contain enum values or variables. Also, lists and objects do not contain any location information since they are generated by resolving fields and values rather than parsing a source query.

Variants
NullScalar(S)List(Vec<Value<S>>)Object(Object<S>)
Methods

impl<S> Value<S> where
S: ScalarValue,

pub fn null() -> Self

Construct a null value.

pub fn int(i: i32) -> Self

Deprecated since 0.11:
Use Value::scalar instead

Construct an integer value.

pub fn float(f: f64) -> Self

Deprecated since 0.11:
Use Value::scalar instead

Construct a floating point value.

pub fn string(s: &str) -> Self

Deprecated since 0.11:
Use Value::scalar instead

Construct a string value.

pub fn boolean(b: bool) -> Self

Deprecated since 0.11:
Use Value::scalar instead

Construct a boolean value.

pub fn list(l: Vec<Self>) -> Self

Construct a list value.

pub fn object(o: Object<S>) -> Self

Construct an object value.

pub fn scalar<T>(s: T) -> Self where
T: Into<S>,

Construct a scalar value

pub fn is_null(&self) -> bool

Does this value represent null?

pub fn as_scalar_value<'a, T>(&'a self) -> Option<&'a T> where
&'a S: Into<Option<&'a T>>,

View the underlying scalar value if present

pub fn as_float_value(&self) -> Option<f64> where
&'a S: ScalarRefValue<'a>,

Deprecated since 0.11:
Use Value::as_scalar_value instead

View the underlying float value, if present.

pub fn as_object_value(&self) -> Option<&Object<S>>

View the underlying object value, if present.

pub fn as_mut_object_value(&mut self) -> Option<&mut Object<S>>

Mutable view into the underlying object value, if present.

pub fn as_list_value(&self) -> Option<&Vec<Self>>

View the underlying list value, if present.

pub fn as_scalar(&self) -> Option<&S>

View the underlying scalar value, if present

pub fn as_string_value<'a>(&'a self) -> Option<&'a str> where
Option<&'a String>: From<&'a S>,

Deprecated since 0.11:
Use Value::as_scalar_value instead

View the underlying string value, if present.