Struct juniper::Executor

pub struct Executor<'a, CtxT, S = DefaultScalarValue> where
    CtxT: 'a,
    S: 'a,  { /* fields omitted */ }

Query execution engine

The executor helps drive the query execution in a schema. It keeps track of the current field stack, context, variables, and errors.

Methods

impl<'a, CtxT, S> Executor<'a, CtxT, S> where
S: ScalarValue,

pub fn resolve_with_ctx<NewCtxT, T>(
&self,
info: &T::TypeInfo,
value: &T
) -> ExecutionResult<S> where
NewCtxT: FromContext<CtxT>,
T: GraphQLType<S, Context = NewCtxT>,

Resolve a single arbitrary value, mapping the context to a new type

pub fn resolve<T>(&self, info: &T::TypeInfo, value: &T) -> ExecutionResult<S> where
T: GraphQLType<S, Context = CtxT>,

Resolve a single arbitrary value into an ExecutionResult

pub async fn resolve_async<'_, '_, '_, T>(
&'_ self,
info: &'_ T::TypeInfo,
value: &'_ T
) -> ExecutionResult<S> where
T: GraphQLTypeAsync<S, Context = CtxT> + Send + Sync,
T::TypeInfo: Send + Sync,
CtxT: Send + Sync,
S: Send + Sync,

Resolve a single arbitrary value into an ExecutionResult

pub async fn resolve_with_ctx_async<'_, '_, '_, NewCtxT, T>(
&'_ self,
info: &'_ T::TypeInfo,
value: &'_ T
) -> ExecutionResult<S> where
T: GraphQLTypeAsync<S, Context = NewCtxT> + Send + Sync,
T::TypeInfo: Send + Sync,
S: Send + Sync,
NewCtxT: FromContext<CtxT> + Send + Sync,

Resolve a single arbitrary value, mapping the context to a new type

pub fn resolve_into_value<T>(&self, info: &T::TypeInfo, value: &T) -> Value<S> where
T: GraphQLType<S, Context = CtxT>,

Resolve a single arbitrary value into a return value

If the field fails to resolve, null will be returned.

pub async fn resolve_into_value_async<'_, '_, '_, T>(
&'_ self,
info: &'_ T::TypeInfo,
value: &'_ T
) -> Value<S> where
T: GraphQLTypeAsync<S, Context = CtxT> + Send + Sync,
T::TypeInfo: Send + Sync,
CtxT: Send + Sync,
S: Send + Sync,

Resolve a single arbitrary value into a return value

If the field fails to resolve, null will be returned.

pub fn replaced_context<'b, NewCtxT>(
&'b self,
ctx: &'b NewCtxT
) -> Executor<'b, NewCtxT, S>

Derive a new executor by replacing the context

This can be used to connect different types, e.g. from different Rust libraries, that require different context types.

pub fn context(&self) -> &'a CtxT

Access the current context

You usually provide the context when calling the top-level execute function, or using the context factory in the Iron integration.

pub fn schema(&self) -> &'a SchemaType<S>

The currently executing schema

pub fn location(&self) -> &SourcePosition

The current location of the executor

pub fn push_error(&self, error: FieldError<S>)

Add an error to the execution engine at the current executor location

pub fn push_error_at(&self, error: FieldError<S>, location: SourcePosition)

Add an error to the execution engine at a specific location

pub fn look_ahead(&'a self) -> LookAheadSelection<'a, S>

Construct a lookahead selection for the current selection.

This allows seeing the whole selection and perform operations affecting the children.