Struct juniper::Object

pub struct Object { /* fields omitted */ }

A Object value

Methods

impl Object

pub fn with_capacity(size: usize) -> Self

Create a new Object value with a fixed number of preallocated slots for field-value pairs

pub fn add_field<K>(&mut self, k: K, value: Value) -> Option<Value> where
K: Into<String>,
&'a str: PartialEq<K>,

Add a new field with a value

If there is already a field with the same name the old value is returned

pub fn contains_field<K>(&self, f: K) -> bool where
&'a str: PartialEq<K>,

Check if the object already contains a field with the given name

pub fn iter(&self) -> FieldIter

Get a iterator over all field value pairs

This method returns a iterator over &'a (String, Value)

pub fn iter_mut(&mut self) -> FieldIterMut

Get a iterator over all mutable field value pairs

This method returns a iterator over &mut 'a (String, Value)

pub fn field_count(&self) -> usize

Get the current number of fields

pub fn get_field_value<K>(&self, key: K) -> Option<&Value> where
&'a str: PartialEq<K>,

Get the value for a given field